+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## _(in-progress)_
 * New: Add support for third-party plugins: Advanced Custom Fields, Elementor
 * New: Add filter `c2c_text_replace_filter_priority` for filtering hook priority for text replacement handler
+* New: Add new setting to allow control over when text replacements are handled early or late in text processing process
 * New: Add filter `c2c_text_replace_third_party_filters` for filtering third party filters
 * Fix: Ensure the lack of any defined replacements doesn't remove zeroes from text
 * Change: Alter handling of `replace_once` value to ensure a valid value is used as arg for `preg_replace()`